refactor: extract recorder project hook
