+++ src/components/recorder/index.tsx
@@ -1,6 +1,5 @@
-import { useMutation, useQuery } from "@tanstack/react-query";
-import { useEffect, useRef, useState, useSyncExternalStore } from "react";
-import { toast } from "sonner";
+import { useMutation } from "@tanstack/react-query";
+import { useEffect, useState, useSyncExternalStore } from "react";
 import { useWindowEvent } from "../../hooks/use-window-event";
 import { resolveAudioFiles } from "../../lib/audio-files";
 import { buildExportFileName, downloadBlob } from "../../lib/export-utils";
@@ -8,7 +7,6 @@ import {
   isShortcutTextInputTarget,
   matchKeyboardEvent,
 } from "../../lib/keyboard";
-import { recorderProjectStorage } from "../../lib/recorder/project-storage";
 import {
   RecorderRuntime,
   RecorderRuntimeState,
@@ -17,7 +15,7 @@ import { formatTimeWithMilliseconds } from "../../lib/time-format";
 import { encodeWav } from "../../lib/wav";
 import { parseTimeSignature } from "../../types";
 import { Dialog } from "../ui/dialog";
-import { RecorderHeader, type SaveStatus } from "./recorder-header";
+import { RecorderHeader } from "./recorder-header";
 import { InputSetup } from "./recorder-input";
 import {
   TakeTimelineLane,
@@ -30,6 +28,7 @@ import {
   TrackRow,
 } from "./recorder-tracks";
 import { useRecorderInput } from "./use-recorder-input";
+import { useRecorderProject } from "./use-recorder-project";
 import { useRecorderTimeline } from "./use-recorder-timeline";
 
 export function Recorder({ projectId }: { projectId: string }) {
@@ -461,78 +460,5 @@ function useRecorderClipSelection({
     },
     select: setSelected,
     selected,
-  };
-}
-
-function useRecorderProject({
-  projectId,
-  runtime,
-}: {
-  projectId: string;
-  runtime: RecorderRuntime;
-}) {
-  const [dirty, setDirty] = useState(false);
-  const revisionRef = useRef(0);
-
-  const projectQuery = useQuery({
-    queryKey: ["recorder-project", projectId],
-    retry: false,
-    staleTime: Infinity,
-    queryFn: async () => {
-      try {
-        const project = await recorderProjectStorage.load(projectId);
-        runtime.deserializeProject(project);
-        return true;
-      } catch (error) {
-        toast.error(error instanceof Error ? error.message : "Unknown error");
-        throw error;
-      }
-    },
-  });
-
-  const saveMutation = useMutation({
-    mutationFn: async () => {
-      const revision = revisionRef.current;
-      await recorderProjectStorage.save({
-        id: projectId,
-        content: runtime.serializeProject(),
-      });
-      return revision;
-    },
-    onSuccess: (savedRevision) => {
-      setDirty(revisionRef.current !== savedRevision);
-    },
-  });
-
-  useEffect(() => {
-    if (!projectQuery.isSuccess) {
-      return;
-    }
-    return runtime.subscribePersistableState(() => {
-      revisionRef.current += 1;
-      setDirty(true);
-    });
-  }, [projectQuery.isSuccess, runtime]);
-
-  useWindowEvent("beforeunload", (event) => {
-    if (dirty) {
-      event.preventDefault();
-    }
-  });
-
-  const saveStatus: SaveStatus = saveMutation.isError
-    ? "error"
-    : saveMutation.isPending
-      ? "saving"
-      : dirty
-        ? "unsaved"
-        : "saved";
-  return {
-    dirty,
-    error: projectQuery.error ?? saveMutation.error,
-    ready: projectQuery.isSuccess || projectQuery.isError,
-    save: saveMutation.mutate,
-    saveStatus,
-    saving: saveMutation.isPending,
   };
 }
